pyproject.toml
[tool.poetry]
name = "projects"
version = "0.1.0"
description = ""
authors = ["MyNameIsClown <vc.business.oo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1"
djangorestframework = "3.15.2"
django-cors-headers = "^4.5.0"
psycopg2 = "^2.9.10"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
